task 2 : Changed type Double-> Int, Test dataSet added

+++ hw2/src/main/java/com/chilik1020/hw2/model/entity/Result.kt
@@ -2,4 +2,4 @@ package com.chilik1020.hw2.model.entity
 
 import java.io.Serializable
 
-data class Result(val average: Double, val sum: Double, val division: Double) : Serializable
+data class Result(val average: Double, val sum: Int, val division: Double) : Serializable
